Merged PR 691: Add Mini Expansion Panel and Accordion Group component
The mini expansion panel component is a stripped down expansion panel for use in Help sidebar.  The accordion group component is used to wrap multiple panels that should only have one open at a time.

The requirements were that only one panel stay open at a time, and that a panel could be set to 'open' by default.

NOTE: some of these files are for the documentation site (the documentation was created all at once)