Merged PR 1072: Convert radio button to SVG
The radio buttons have been plagued by what I call the 'googly eye problem', where the inner circle of the radio button is off-center due to how various browsers interpret fractions of pixels.  By converting them to SVG icons the browser has a better understanding of the shapes in relation to each other, which fixes the problem.

One thing I learned about the icon font builder, however, is that it changes which icons are assigned to which font glyph when you add a new icon and re-build (assigned in alphabetical order, so inserting one in the middle bumps the preceding icons down the ranks).  This isn't a problem when using the icon through the css class (which doesn't change), but it does pose a problem for instances where the icon was referenced by their unicode number (which is dynamic).  To fix this problem I altered the mixins that are available in the icon font scss file, and use the `@extend .inner-rivet-icon-check !optional;` pattern.  The `!optional` flag is required, because the icon mixin is not directly available until the stylesheets are compiled, and the build script has a fit about it without that flag.

Checked for compatibility in IE, Edge, Firefox, Chrome
- Edge ignores hover state, but it did for the old ones too, so it's not a new problem.