Merged PR 1130: Update rivet tooltip init logic
We discovered that tooltips weren't working on material inputs because the material content was being transformed during the angular lifecycle, so the directive got screwed up on content that was basically being transcluded.  Long story short, using AfterViewInit fixes the problem.

Also, there was a pointer-events css style that was preventing tooltips from working on mat-input labels, so I made sure to allow pointer events on anything that has a rvtTooltip.

Also also: mat-input-labels have overflow hidden, which is not very cash money, so I overrode that as well so you could see tooltips attached to them.