Merged PR 1392: Upgrade Rivet 4 to Angular 9
CALLING ALL UX AMBASSADORS

Rivet 4.x is intended to go out with NaviPlan Essentials, and as such will need to be running Angular 9.  This is that upgrade.
Rivet v3.5.9 has already been branched for Admin and Guided Retirement, so those projects will not be required to upgrade Rivet.
Talking to Cirese this should also not be a disruption of any AuthServer work, as it does not consume Rivet directly.

NOTE: the package.json references the latest version of Rivet in the Azure NPM feed, so you will not see the update locally if you run the documentation site.  Change 'latest' to 'file:dist/rivet-style' to link the package build from the build_lib script.