Merged PR 2087: Tooltip global position option
Admin has been running into problems where a tooltip that was placed inline with it's anchor element was being cut off by an overflow: hidden on a parent element.  This PR solves that problem by offering an option that places the tooltip at the body level instead of inline with the anchor.

Pros: doesn't get cut off by overflow, z-index is always on top.
Cons: doesn't follow element when resizing screen or when element shifts.