Merged PR 2141: Repo cleanup
For whatever reason there was a 'RIvet' folder at the top level of the repo that has old copies of duplicated files.  My assumption is that somebody cloned the repo in a weird way and didn't notice right away.

How did it get there?  I don't know.
When did it happen?  Sometime around the migration to Git.
Are any of these files used?  Nope, everything builds and runs without it.