Merged PR 4030: BrowserModule duplicate fix
It has been discovered that since v4.7.1 there was an issue introduced where the BrowserAnimationsModule imported a duplicate BrowserModule in Essentials on the Client Facts page.  This has been rectified by relying on the CommonModules instead, which has been tested in an alpha package version to successfully fix the issue in Essentials, and has been tested to work with the drawer component that previously used the animation module.