Merged PR 4860: Update rivet tiles to allow for non-selectable tiles
There are a few designs for Essentials that will be including tiles that look and work in a similar manner to original tiles, but users will be unable to select the tile.

If no select event is passed into the tile it will be disabled by default and will not have the hover effect.
If the tile is selectable, then it has the option to be disabled in the config and then it will get the disabled styling (dashed border, lighter color)
Edit and delete will still work on both types of tiles if they are passed in.