Merged PR 6229: Fix build script
For whatever reason in the VM update that Joe pushed recently, cpx (the npm script that copies icons and fonts into the dist folder) started failing copying the things it was supposed to copy.  After tinkering a bit it seems the problem was that cpx was expecting the directory path strings to be wrapped in double quotes.  A test run of the pipline on this branch seems to agree, as the copy step was successful with this fix.